chore(firstRepeatWord/consolidate): Test and Problem Consolidated
All consolidated.
