+++ FirstRepeatedWordInASentence/Problem/FirstRepeatWordFinder.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace FirstRepeatedWordInASentence
+namespace FirstRepeatedWordInASentence.Problem
 {
     public class FirstRepeatWordFinder
     {
