+++ FirstRepeatedWordInASentence/packages.config
